Testing Improvement

+++ Prism/src/test/java/me/botsko/prism/testHelpers/TestHelper.java
@@ -3,11 +3,14 @@ package me.botsko.prism.testHelpers;
 import be.seeseemelk.mockbukkit.MockBukkit;
 import be.seeseemelk.mockbukkit.ServerMock;
 import me.botsko.prism.Prism;
+import me.botsko.prism.database.PrismDataSource;
+import org.bstats.bukkit.Metrics;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 
 /**
@@ -19,6 +22,7 @@ public class TestHelper {
     public static ServerMock setup() {
         ServerMock server = MockBukkit.getOrCreateMock();
         server.addSimpleWorld("Normal");
+        Metrics metrics = null;
         MockBukkit.load(Prism.class);
         server.getScheduler().waitAsyncTasksFinished();
         server.getScheduler().performTicks(100);
@@ -26,28 +30,32 @@ public class TestHelper {
     }
 
     public static void shutdown() {
+        PrismDataSource dataSource = Prism.getPrismDataSource();
+        if(dataSource != null) {
+            Collection<String> sql = new ArrayList<>();
+            sql.add("SET FOREIGN_KEY_CHECKS=0;");
+            sql.add("DROP TABLE IF EXISTS prism_actions;");
+            sql.add("DROP TABLE IF EXISTS prism_data;");
+            sql.add("DROP TABLE IF EXISTS prism_data_extra;");
+            sql.add("DROP TABLE IF EXISTS prism_meta;");
+            sql.add("DROP TABLE IF EXISTS prism_players;");
+            sql.add("DROP TABLE IF EXISTS prism_worlds;");
+            sql.add("DROP TABLE IF EXISTS prism_id_map;");
+            sql.add("DROP TABLE IF EXISTS prism_players;");
+            sql.forEach(s -> {
+
+                try (
+                        Connection conn = Prism.getPrismDataSource()
+                                .getDataSource().getConnection()
+                ) {
+                    PreparedStatement statement = conn.prepareStatement(s);
+                    statement.execute();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            });
+        }
         MockBukkit.getMock().getPluginManager().disablePlugins();
         MockBukkit.unmock();
-        List<String> sql = new ArrayList<>();
-        sql.add("SET FOREIGN_KEY_CHECKS=0;");
-        sql.add("DROP TABLE IF EXISTS prism_actions;");
-        sql.add("DROP TABLE IF EXISTS prism_data;");
-        sql.add("DROP TABLE IF EXISTS prism_data_extra;");
-        sql.add("DROP TABLE IF EXISTS prism_meta;");
-        sql.add("DROP TABLE IF EXISTS prism_players;");
-        sql.add("DROP TABLE IF EXISTS prism_worlds;");
-        sql.add("DROP TABLE IF EXISTS prism_id_map;");
-        sql.add("DROP TABLE IF EXISTS prism_players;");
-        sql.forEach(s -> {
-            try (
-                    Connection conn = Prism.getPrismDataSource()
-                            .getDataSource().getConnection()
-            ) {
-                PreparedStatement statement = conn.prepareStatement(s);
-                statement.execute();
-            } catch (SQLException e) {
-                e.printStackTrace();
-            }
-        });
     }
 }
